Mouse movement, Rotation fixes, misc fixes

+++ src/A2/scripts/PlayerControllerScript.h
@@ -12,7 +12,7 @@
 class PlayerControllerScript : public ScriptComponent {
 private:
     int lastShoot = -1;
-    int lastParticleEmit = -1;
+    float lastParticleEmit = -1;
     float velocity = 0;
 
     CameraComponent* camera;
